AudioFrame: fix get/set function of fps
The type is double.

+++ src/AvTranscoder/frame/AudioFrame.hpp
@@ -31,7 +31,7 @@ public:
 
 	size_t getSampleRate() const { return _sampleRate; }
 	size_t getChannels() const { return _channels; }
-	size_t getFps() const { return _fps; }
+	double getFps() const { return _fps; }
 	AVSampleFormat getSampleFormat() const { return _sampleFormat; }
 	std::string getSampleFormatName() const
 	{
@@ -53,7 +53,7 @@ public:
 	
 	void setSampleRate( const size_t sampleRate ) { _sampleRate = sampleRate; }
 	void setChannels( const size_t channels ) { _channels = channels; }
-	void setFps( const size_t fps ) { _fps = fps; }
+	void setFps( const double fps ) { _fps = fps; }
 	void setSampleFormat( const std::string& sampleFormatName ) { _sampleFormat = av_get_sample_fmt( sampleFormatName.c_str() ); }
 	void setSampleFormat( const AVSampleFormat sampleFormat ) { _sampleFormat = sampleFormat; }
